@@ -1,14 +1,15 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import store from './store';
 
 import App from './container/App';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(
-	<Provider store={store}>
-		<App />
-	</Provider>,
-	document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
+root.render(
+        <Provider store={store}>
+                <App />
+        </Provider>
+);
 registerServiceWorker();
